% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{angle}
\alias{angle}
\title{Calculates angle of 2-dimensional vectors.}
\usage{
angle(x, y)
}
\arguments{
\item{x}{vector 1.}

\item{y}{vector 2.}
}
\value{
angle in radian.
}
\description{
This function calculates the angle of x to y in anticlockwise.
}
\examples{
# a example of orthogonal vectors

angle(c(1, 0), c(0, 1))
# = 1 / 2 * pi

angle(c(0, 1), c(1, 0))
# = 3 / 2 * pi

}
