% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{databar}
\alias{databar}
\title{Plot databars for matrix visualization}
\usage{
databar(X, grid = T)
}
\arguments{
\item{X}{target matrix.}

\item{grid}{if \code{T}, drawing grid lines that separete each cells of matrix.}
}
\description{
Plot databars for matrix visualization
}
\examples{
databar(matrix(1:9, 3, 3))

}
