% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{debugText}
\alias{debugText}
\title{Prints variables or expressions  for debug.}
\usage{
debugText(...)
}
\arguments{
\item{...}{printed objects.}
}
\description{
Prints variables or expressions  for debug.
}
\examples{
　
f <- function(x,y){
  debugText(x^y,y)
  return(x^y+y)
}
f(4,6)
}
