% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{dir.create.deep}
\alias{dir.create.deep}
\title{Checks and creates a unknown directory when it is not exist.}
\usage{
dir.create.deep(pass)
}
\arguments{
\item{pass}{target directory pass.}
}
\description{
Checks and creates a unknown directory when it is not exist.
}
