% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{euclidean}
\alias{euclidean}
\title{Calculates euclidean distance of vectors.}
\usage{
euclidean(x, y)
}
\arguments{
\item{x}{a vector.}

\item{y}{a vector.}
}
\value{
euclidean distance between \code{x} and \code{y}.
}
\description{
Calculates euclidean distance of vectors.
}
