% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{evalTimer}
\alias{evalTimer}
\title{Measure the elapsed time of \code{eval} and returns the result.}
\usage{
evalTimer(eval, keep = F)
}
\arguments{
\item{eval}{evalated contents.}

\item{keep}{if \code{T}, returns evalated object with elapsed time.}
}
\value{
evalated object.
}
\description{
Measure the elapsed time of \code{eval} and returns the result.
}
\examples{
runif.sum <- evalTimer(sum(runif(5000)))
}
