% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{\%.\%}
\alias{\%.\%}
\title{magrittr operator with giving nothing to first argment}
\usage{
lhs \%.\% rhs
}
\arguments{
\item{lhs}{a value or the magrittr placeholder.}

\item{rhs}{a function call using the magrittr semantics.}
}
\description{
magrittr operator with giving nothing to first argment
}
\examples{
# this is equivalent to `2 \%>\% {rep(5, .)} \%>\% sqrt`
2 \%.\% rep(5, .) \%>\% sqrt

}
