% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{\%T.\%}
\alias{\%T.\%}
\title{magrittr tee operator with giving nothing to first argment}
\usage{
lhs \%T.\% rhs
}
\arguments{
\item{lhs}{a value or the magrittr placeholder.}

\item{rhs}{a function call using the magrittr semantics.}
}
\description{
magrittr tee operator with giving nothing to first argment
}
\examples{
require(magrittr)

#
# this is equivalent to
# 1:2 \%T>\% {plot(0:3, xlim = .)} \%>\% sqrt
#
1:2 \%T.\% plot(0:3, xlim = .) \%>\% sqrt

}
