% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{image2}
\alias{image2}
\title{Improved \code{image} function for matrix visualization}
\usage{
image2(X, col.rainbow = F, col.gray0 = F)
}
\arguments{
\item{X}{target matrix.}

\item{col.rainbow}{if \code{T}, uses rainbow color.}

\item{col.gray0}{if \code{T}, the color of cells where values are 0 is set to gray.}
}
\description{
This function show \code{matrix} with the direction likes \code{print}.
In addition, adds the color bar and labels of axis from dimnames.
}
