% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is.prob}
\alias{is.prob}
\title{Asserts \code{x} whether probabic number or not.}
\usage{
is.prob(x)
}
\arguments{
\item{x}{tested object..}
}
\description{
Asserts \code{x} whether probabic number or not.
}
