% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is.range}
\alias{is.range}
\title{Asserts \code{x} whether numeric range or not.}
\usage{
is.range(x)
}
\arguments{
\item{x}{object to test.}
}
\description{
Asserts \code{x} whether numeric range or not.
}
