% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{is.todayDirectory}
\alias{is.todayDirectory}
\title{Checks if current directory is today directory or not.}
\usage{
is.todayDirectory()
}
\description{
Checks if current directory is today directory or not.
}
