% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{keywait}
\alias{keywait}
\title{Waits until enter key is pressd.}
\usage{
keywait()
}
\description{
Waits until enter key is pressd.
}
\examples{
　
f <- function(n){
  mat <- matrix(runif(n**2,n,n))
  plot(mat); keywait()
  image(mat)
}
f(3)
}
