% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{matlist}
\alias{matlist}
\title{Converts lists or vectors that nested by list to matrix.}
\usage{
matlist(lst, nrow = length(lst), ncol = length(lst[[nrow]]),
  rname = names(lst), cname = names(lst[[nrow]]))
}
\arguments{
\item{lst}{lists or vectors nested by list.}

\item{nrow}{number of rows.}

\item{ncol}{number of cols.}

\item{rname}{rownames of matrix after converted.}

\item{cname}{colnames of matrix after converted.}
}
\value{
converted matrix.
}
\description{
Converts lists or vectors that nested by list to matrix.
}
\examples{
lst <- lapply(1:5,rep,3)
matlist(lst)

}
