% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{max2nd}
\alias{max2nd}
\title{Calculate second highest value in vector.}
\usage{
max2nd(x)
}
\arguments{
\item{x}{target vector.}
}
\value{
second highest value.
}
\description{
Calculate second highest value in vector.
}
