% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{memory_summary}
\alias{memory_summary}
\title{Summarize object size in global environment.}
\usage{
memory_summary()
}
\value{
a tibble that contains summarizing result.
}
\description{
Summarize object size in global environment.
}
\references{
this function was impressed by \code{pryr:::print.bytes()}
authored by Hadly Wickman.
}
