% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{min2nd}
\alias{min2nd}
\title{Calculate second lowest value in vector.}
\usage{
min2nd(x)
}
\arguments{
\item{x}{target vector.}
}
\value{
second lowest value.
}
\description{
Calculate second lowest value in vector.
}
