% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{order.col}
\alias{order.col}
\title{Orders matrix in each rows.}
\usage{
order.col(X, decreasing = F, ignore0 = F)
}
\arguments{
\item{X}{target matrix.}

\item{decreasing}{if \code{T}, orders in decreasing.}

\item{ignore0}{special option when \code{X} is distance matrix.}
}
\value{
matrix with orderd each rows.
}
\description{
Orders matrix in each rows.
}
