% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{order2}
\alias{order2}
\title{\code{order} function regards a duplication value as the same ranking.}
\usage{
order2(x, decreasing = F)
}
\arguments{
\item{x}{input vector.}

\item{decreasing}{if \code{TRUE}, proceses in decreasing.}
}
\value{
order of \code{x}.
}
\description{
\code{order} function regards a duplication value as the same ranking.
}
