% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{padding.matrix}
\alias{padding.matrix}
\title{Padding (expanding rows and cols) for \code{matrix}.}
\usage{
padding.matrix(X, size = 1, replace = NULL)
}
\arguments{
\item{X}{target matrix.}

\item{size}{expanding size.}

\item{replace}{values of expanded elements. if NULL, copies border values of the matrix.}
}
\value{
padding matrix.
}
\description{
Padding (expanding rows and cols) for \code{matrix}.
}
