% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magrittr.R
\name{passign}
\alias{passign}
\title{assign function for pipe stream}
\usage{
passign(value, x, function_list = list(), envir = globalenv())
}
\arguments{
\item{value}{a value to be assigned to \code{x}.}

\item{x}{a variable name.}

\item{function_list}{a list of functions.}

\item{envir}{the environment to use.}
}
\value{
\code{value} with applied function_list.
}
\description{
assign function for pipe stream
}
\examples{
x <- 1:4
x.sum <- x \%>\% passign(x.sqrt, function_list = list(sqrt)) \%>\% sum

print(x.sum)
print(x.sqrt)

}
