% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.lines}
\alias{plot.lines}
\title{Plot matlix as a set of lines}
\usage{
\method{plot}{lines}(X, legend.locale = "topleft")
}
\arguments{
\item{X}{target matrix.}

\item{legend.locale}{location of legend, for example, "topleft" or "bottomright"}
}
\description{
Plot matlix as a set of lines
}
\examples{

x <- seq(0, 2*pi, length.out = 100)
X <- rbind(sin(x), cos(x), sin(x) + cos(x))
rownames(X) <- c("sin(x)", "cos(x)", "sin(x) + cos(x)")

plot.lines(X, "topright")

}
