% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rnotebook.R
\name{preparePrintify}
\alias{preparePrintify}
\title{Prepare the process that enables to print out your R-notebook document more beautifully}
\usage{
preparePrintify()
}
\description{
enable to background color of code chunks and remove "hide" button.
}
