% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{print.noattr}
\alias{print.noattr}
\title{Print variable without inconsiderable attributes.}
\usage{
\method{print}{noattr}(x, keeps = c("names", "row.names", "class", "dim",
  "dimnames"), ...)
}
\arguments{
\item{x}{printed variable.}

\item{keeps}{names of the considerable attributes.}

\item{...}{arbitrary argments for print.}
}
\description{
Print variable without inconsiderable attributes.
}
