% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{quadrature}
\alias{quadrature}
\title{Integrates discrete vector by quadrature.}
\usage{
quadrature(x)
}
\arguments{
\item{x}{target density object calculated by function \code{density()}.}
}
\value{
integration of \code{x}.
}
\description{
Integrates discrete vector by quadrature.
}
\examples{
n <- rnorm(100)
dens <- density(n)
quadrature(dens)

}
