% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rank.matrix}
\alias{rank.matrix}
\title{Calculates rank of each elements of matrix.}
\usage{
rank.matrix(X)
}
\arguments{
\item{X}{target matrix.}
}
\value{
matrix presents rank.
}
\description{
Calculates rank of each elements of matrix.
}
