% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{replaceNA}
\alias{replaceNA}
\title{Replaces NA from matrix with maintaining the dimension.}
\usage{
replaceNA(X, replace = 0)
}
\arguments{
\item{X}{target matrix.}

\item{replace}{value to be replaced for NA.}
}
\value{
NA removed matrix.
}
\description{
Replaces NA from matrix with maintaining the dimension.
}
