% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rowMinus}
\alias{rowMinus}
\title{Calculates \code{X - a} with row first.}
\usage{
rowMinus(X, a)
}
\arguments{
\item{X}{Minused matrix.}

\item{a}{Minuses vector.}
}
\value{
\code{X - a} in row first.
}
\description{
R's default calculation of \code{X - a} is with col first likes \code{X[,1] - a}, \code{X[,2] - a}, ....
This function implements calclation with row first likes \code{X[1,] - a}, \code{X[1,] - a}, ....
}
