% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rowNorm}
\alias{rowNorm}
\title{Calculates norm for each row of \code{X}}
\usage{
rowNorm(X)
}
\arguments{
\item{X}{input numeric matrix.}
}
\value{
norm for each row of \code{X}.
}
\description{
Calculates norm for each row of \code{X}
}
