% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{rowTimes}
\alias{rowTimes}
\title{Calculates \code{X - a}}
\usage{
rowTimes(X, a)
}
\arguments{
\item{X}{Timed matrix.}

\item{a}{Times vector.}
}
\value{
\code{X \* a} in row first.
}
\description{
R's default calculation of \code{X \* a} is in col first likes \code{X[,1] \* a}, \code{X[,2] \* a}, ....
This function implements calclation in row first likes \code{X[1,] \* a}, \code{X[1,] \* a}, ....
}
