% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{runifN}
\alias{runifN}
\title{Generates random natural numbers}
\usage{
runifN(n, min = 1, max = 10)
}
\arguments{
\item{n}{number of the random numbers.}

\item{min}{min of random numbers.}

\item{max}{max of random numbers.}
}
\value{
random natural numbers.
}
\description{
Generates random natural numbers
}
