% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{savedDir}
\alias{savedDir}
\title{Retruns or Sets default save directory.}
\usage{
savedDir(pass)
}
\arguments{
\item{pass}{save directory pass if you want to change this.}
}
\value{
save directory
}
\description{
When \code{saveDir} called without \code{pass}, returns default save directory which is
contained package directory.
When with \code{pass}, sets default save directory as pass.
}
