% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{saveEnvironment}
\alias{saveEnvironment}
\title{Save global environment.}
\usage{
saveEnvironment()
}
\description{
Save global environment as non named file '.RData' without closing RStudio.
Directory to save is set project directory.
}
