% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{saveggpng}
\alias{saveggpng}
\title{Save plot generated by ggplot2.}
\usage{
saveggpng(name, current = F, scale = c(450, 500), size = 20)
}
\arguments{
\item{name}{names of the saved image file without ".png".}

\item{current}{force create the sub directory even if current directory is not today directory.}

\item{scale}{size of saved image.}

\item{size}{text size.}
}
\description{
Changes text size of figure to larger to improve visibility.
Then save plot with \code{\link[=savepng]{savepng()}}
}
\seealso{
\code{\link[=savepng]{savepng()}}
}
