% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{savepng}
\alias{savepng}
\title{Saves plot to today directory.}
\usage{
savepng(name, current = F, scale = c(450, 500))
}
\arguments{
\item{name}{names of the saved image file without ".png".}

\item{current}{force create the sub directory even if current directory is not today directory.}

\item{scale}{size of saved image.}
}
\description{
Saves plot to today directory.
}
\examples{
plot(sin, xlim = c(0,2*pi))
#todaywd()
#savepng("sin curve")

}
