% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{savepngTitle}
\alias{savepngTitle}
\title{Saves plot with adding title on the figure.}
\usage{
savepngTitle(name, current = F, scale = c(450, 500))
}
\arguments{
\item{name}{title of the plot.}

\item{current}{force create the sub directory even if current directory is not today directory.}

\item{scale}{size of saved image.}
}
\description{
Saves plot with adding title on the figure.
}
