% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{setMargin}
\alias{setMargin}
\title{Changes margins for plot}
\usage{
setMargin(top.4 = 4, left.4 = 4, bottom.5 = 5, right.2 = 2)
}
\arguments{
\item{top.4}{maring of top. default is 4.}

\item{left.4}{margin of left. defualt is 4.}

\item{bottom.5}{margin of bottom. default is 5.}

\item{right.2}{margin of right. default is 2.}
}
\description{
Changes margins for plot
}
