% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{softwriteEllipsis}
\alias{softwriteEllipsis}
\title{Overwrite some desirable arguments to \code{...} softly.}
\usage{
softwriteEllipsis(..., append = NULL, warn = F)
}
\arguments{
\item{...}{fixed and given arguments.}

\item{append}{argments list to be appended.}

\item{warn}{if \code{T}, warnings when fixed and given argments conflict.}
}
\value{
overwrited ellipsis.
}
\description{
This function is minor modificated version of \code{\link{overwriteEllipsis}}.
The only difference is that don't overwrite conflict argments.
}
\examples{
# plot function with xlim = c(0, 2 * pi) and col = 'red' defaultly
plot2pi <- function(x, ...){
  elp <- overwriteEllipsis(..., x = x, xlim = c(0, 2 * pi))
  elp <- softwriteEllipsis(..., append = elp, col = 'red')
  do.call(plot, elp)
}

# this example is evaluated as col = 'red'
plot2pi(sin)

# this example is evaluated as col = 'blue'
plot2pi(sin, col = 'blue')

}
\seealso{
\code{\link{overwriteEllipsis}}
}
