% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{sort_col}
\alias{sort_col}
\title{Sorts matrix in each rows.}
\usage{
sort_col(X, decreasing = F, ignore0 = F)
}
\arguments{
\item{X}{target matrix.}

\item{decreasing}{if \code{T}, sorts in decreasing.}

\item{ignore0}{special option when \code{X} is distance matrix.}
}
\value{
matrix with sorted each rows.
}
\description{
Sorts matrix in each rows.
}
