% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_output.R
\name{subwd}
\alias{subwd}
\title{Creates sub directory for saving a number of images.}
\usage{
subwd(name, current = F)
}
\arguments{
\item{name}{name of the sub directory.}

\item{current}{force create the sub directory even if current directory is not today directory.}
}
\description{
Creates sub directory for saving a number of images.
}
