% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{which.max.ind}
\alias{which.max.ind}
\title{Searches index of maximum element from matrix.}
\usage{
which.max.ind(X)
}
\arguments{
\item{X}{target matrix.}
}
\value{
index of maximum element.
}
\description{
Searches index of maximum element from matrix.
}
