% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation.R
\name{which.min.ind}
\alias{which.min.ind}
\title{Searches index of minimum element from matrix.}
\usage{
which.min.ind(X)
}
\arguments{
\item{X}{target matrix.}
}
\value{
index of minimum element.
}
\description{
Searches index of minimum element from matrix.
}
