% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{overwriteEllipsis}
\alias{overwriteEllipsis}
\title{Overwrite some desirable arguments to the valiability arguments \code{...}.}
\usage{
overwriteEllipsis(..., warn = T)
}
\arguments{
\item{...}{fixed and given arguments.}

\item{warn}{if \code{T}, warnings when fixed and given argments conflict.}
}
\value{
overwrited ellipsis.
}
\description{
When ellipsis arguments is given from the parents function call as \code{...},
frequently we want to overwrite some arguments as fixed desirable arguments.
This function settles the situation simply.
}
\examples{
plot2pi <- function(x, ...){
  ellipsis <- overwriteEllipsis(..., x = x, xlim = c(0, 2*pi), col = "red")
  do.call(plot, ellipsis)
}

plot2pi(sin)

# this example is warned
plot2pi(sin, main = "sin curve", xlim = c(-pi, pi))

}
