% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{savedDir}
\alias{savedDir}
\title{Retruns or Sets default save directory.}
\usage{
savedDir(pass)
}
\arguments{
\item{pass}{save directory pass if you want to change this.}
}
\value{
save directory
}
\description{
Retruns or Sets default save directory.
}
