% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{savepngTitle}
\alias{savepngTitle}
\title{Saves plot with title.}
\usage{
savepngTitle(name, scale = c(450, 500))
}
\arguments{
\item{name}{title of the plot.}

\item{scale}{size of saved image.}
}
\description{
Saves plot with title.
}
